Split starknet plugin to multiple files.
