+++ crates/cairo-lang-starknet/src/contract.rs
@@ -9,7 +9,8 @@ use cairo_lang_semantic::plugin::DynPluginAuxData;
 use num_bigint::BigUint;
 use sha3::{Digest, Keccak256};
 
-use crate::plugin::{StarkNetContractAuxData, ABI_TRAIT};
+use crate::plugin::aux_data::StarkNetContractAuxData;
+use crate::plugin::consts::ABI_TRAIT;
 
 #[cfg(test)]
 #[path = "contract_test.rs"]
