+++ crates/cairo-lang-starknet/src/plugin/consts.rs
@@ -1,0 +1,11 @@
+pub const VIEW_ATTR: &str = "view";
+pub const EVENT_ATTR: &str = "event";
+pub const ABI_TRAIT: &str = "__abi";
+pub const STORAGE_STRUCT_NAME: &str = "Storage";
+pub const EXTERNAL_MODULE: &str = "__external";
+pub const CONSTRUCTOR_MODULE: &str = "__constructor";
+
+pub(super) const ABI_ATTR: &str = "abi";
+pub(super) const CONTRACT_ATTR: &str = "contract";
+pub(super) const EXTERNAL_ATTR: &str = "external";
+pub(super) const CONSTRUCTOR_ATTR: &str = "constructor";
